+++ test/schema/schema.definition.sql
@@ -765,7 +765,7 @@ SET default_table_access_method = heap;
 
 CREATE TABLE maevsi.event (
     id uuid DEFAULT gen_random_uuid() NOT NULL,
-    address uuid,
+    address_id uuid,
     description text,
     "end" timestamp with time zone,
     guest_count_maximum integer,
@@ -810,10 +810,10 @@ The event''s internal id.';
 
 
 --
--- Name: COLUMN event.address; Type: COMMENT; Schema: maevsi; Owner: postgres
+-- Name: COLUMN event.address_id; Type: COMMENT; Schema: maevsi; Owner: postgres
 --
 
-COMMENT ON COLUMN maevsi.event.address IS 'The event''s physical address.';
+COMMENT ON COLUMN maevsi.event.address_id IS 'Optional reference to the physical address of the event.';
 
 
 --
@@ -1176,30 +1176,39 @@ COMMENT ON FUNCTION maevsi.events_organized() IS 'Add a function that returns al
 --
 
 CREATE FUNCTION maevsi.guest_claim_array() RETURNS uuid[]
-    LANGUAGE plpgsql STABLE STRICT
+    LANGUAGE plpgsql STABLE STRICT SECURITY DEFINER
     AS $$
 DECLARE
   _guest_ids UUID[];
   _guest_ids_unblocked UUID[] := ARRAY[]::UUID[];
-  _guest_id UUID;
 BEGIN
   _guest_ids := string_to_array(replace(btrim(current_setting('jwt.claims.guests', true), '[]'), '"', ''), ',')::UUID[];
 
   IF _guest_ids IS NOT NULL THEN
-    FOREACH _guest_id IN ARRAY _guest_ids
-    LOOP
-      -- omit guests of events created by an account blocked by the current user
-      IF EXISTS (
-	      SELECT 1
-	      FROM maevsi.guest g
-	        JOIN maevsi.event e ON g.event_id = e.id
-	      WHERE g.id = _guest_id AND e.created_by NOT IN (
+    _guest_ids_unblocked := ARRAY (
+      SELECT g.id
+      FROM maevsi.guest g
+        JOIN maevsi.event e ON g.event_id = e.id
+        JOIN maevsi.contact c ON g.contact_id = c.id
+      WHERE g.id = ANY(_guest_ids)
+        AND e.created_by NOT IN (
             SELECT id FROM maevsi_private.account_block_ids()
           )
-	    ) THEN
-        _guest_ids_unblocked := array_append(_guest_ids_unblocked, _guest_id);
-	    END IF;
-    END LOOP;
+        AND (
+          c.created_by NOT IN (
+            SELECT id FROM maevsi_private.account_block_ids()
+          )
+          AND (
+            c.account_id IS NULL
+            OR
+            c.account_id NOT IN (
+              SELECT id FROM maevsi_private.account_block_ids()
+            )
+          )
+        )
+    );
+  ELSE
+    _guest_ids_unblocked := ARRAY[]::UUID[];
   END IF;
   RETURN _guest_ids_unblocked;
 END
@@ -1845,10 +1854,12 @@ CREATE FUNCTION maevsi_private.account_block_ids() RETURNS TABLE(id uuid)
     AS $$
 BEGIN
   RETURN QUERY
+    -- users blocked by the current user
     SELECT blocked_account_id
     FROM maevsi.account_block
     WHERE created_by = maevsi.invoker_account_id()
     UNION ALL
+    -- users who blocked the current user
     SELECT created_by
     FROM maevsi.account_block
     WHERE blocked_account_id = maevsi.invoker_account_id();
@@ -2963,7 +2974,7 @@ Reference to the account that last updated the address.';
 CREATE TABLE maevsi.contact (
     id uuid DEFAULT gen_random_uuid() NOT NULL,
     account_id uuid,
-    address uuid,
+    address_id uuid,
     email_address text,
     email_address_hash text GENERATED ALWAYS AS (md5(lower("substring"(email_address, '\S(?:.*\S)*'::text)))) STORED,
     first_name text,
@@ -3012,10 +3023,10 @@ COMMENT ON COLUMN maevsi.contact.account_id IS 'Optional reference to an associa
 
 
 --
--- Name: COLUMN contact.address; Type: COMMENT; Schema: maevsi; Owner: postgres
+-- Name: COLUMN contact.address_id; Type: COMMENT; Schema: maevsi; Owner: postgres
 --
 
-COMMENT ON COLUMN maevsi.contact.address IS 'Physical address of the contact. Must be between 1 and 300 characters.';
+COMMENT ON COLUMN maevsi.contact.address_id IS 'Optional reference to the physical address of the contact.';
 
 
 --
@@ -3525,7 +3536,7 @@ CREATE VIEW maevsi.guest_flat WITH (security_invoker='true') AS
     guest.feedback_paper AS guest_feedback_paper,
     contact.id AS contact_id,
     contact.account_id AS contact_account_id,
-    contact.address AS contact_address,
+    contact.address_id AS contact_address_id,
     contact.email_address AS contact_email_address,
     contact.email_address_hash AS contact_email_address_hash,
     contact.first_name AS contact_first_name,
@@ -5324,11 +5335,11 @@ ALTER TABLE ONLY maevsi.contact
 
 
 --
--- Name: contact contact_address_fkey; Type: FK CONSTRAINT; Schema: maevsi; Owner: postgres
+-- Name: contact contact_address_id_fkey; Type: FK CONSTRAINT; Schema: maevsi; Owner: postgres
 --
 
 ALTER TABLE ONLY maevsi.contact
-    ADD CONSTRAINT contact_address_fkey FOREIGN KEY (address) REFERENCES maevsi.address(id);
+    ADD CONSTRAINT contact_address_id_fkey FOREIGN KEY (address_id) REFERENCES maevsi.address(id);
 
 
 --
@@ -5340,11 +5351,11 @@ ALTER TABLE ONLY maevsi.contact
 
 
 --
--- Name: event event_address_fkey; Type: FK CONSTRAINT; Schema: maevsi; Owner: postgres
+-- Name: event event_address_id_fkey; Type: FK CONSTRAINT; Schema: maevsi; Owner: postgres
 --
 
 ALTER TABLE ONLY maevsi.event
-    ADD CONSTRAINT event_address_fkey FOREIGN KEY (address) REFERENCES maevsi.address(id);
+    ADD CONSTRAINT event_address_id_fkey FOREIGN KEY (address_id) REFERENCES maevsi.address(id);
 
 
 --